0.2.0 Created new mutations for ProductVariation, VariationDetails. Created new queries for Country, ProductVariation, VariationDetails. Added description field to Product. Changed amount field in ProductVariation to float type. Slightly changed OrderDetails and BasketDetails. Created new DTOs for mutations.

+++ src/main/java/org/example/entity/ProductVariation.java
@@ -20,7 +20,7 @@ public class ProductVariation {
     @OneToOne
     @JoinColumn(name = "image_id")
     private Image variationImage;
-    private double amount;
+    private float amount;
     @OneToMany(mappedBy = "productVariation", fetch = FetchType.LAZY)
     private List<VariationDetails> variationDetails;
     @ManyToOne(fetch = FetchType.LAZY)
